Add share button to website (design by design team).
