+++ stylus/extension/index.styl
@@ -23,6 +23,9 @@
   &:hover
     opacity: 0.85
 
+  span
+    padding-left 15px
+
   img
     position absolute
     top 5px
@@ -81,6 +84,13 @@
     margin-top 10px
     padding 5px 0
     font-size 20px
+    text-align center
+    &.valid-token
+      background-color #00B9F1
+      color white
+    &.invalid-token
+      background-color red
+      color white
 
     i 
       line-height 22px
