+++ views/token-editor-common.pug
@@ -29,7 +29,7 @@
             <span id="hmacsha-text">HMACSHA256</span>(
               base64UrlEncode(header) + "." +
               base64UrlEncode(payload),
-              <input type="text" name="secret" value="secret" title="Weak secret!">
+              <input type="text" name="secret" value="your-256-bit-secret" title="Weak secret!">
             )<span class="is-base64-encoded-label"> <input id="is-base64-encoded" type="checkbox" name="is-base64-encoded"><label for="is-base64-encoded">secret base64 encoded</label></span>
           pre.pre-encode.RS256.
             <span id="rsasha-text">RSASHA256</span>(
@@ -38,5 +38,3 @@
               <textarea rows="4" name="public-key" placeholder="Public Key or Certificate. Enter it in plain text only if you want to verify a token"></textarea>,
               <textarea rows="5" name="private-key" placeholder="Private Key. Enter the it in plain text only if you want to generate a new token. The key never leaves your browser."></textarea>
             )
-
-.validation-status.js-signature
