+++ views/token-editor-validation.pug
@@ -1,0 +1,1 @@
+.validation-status.js-signature
