+++ views/website/token-editor.pug
@@ -11,3 +11,10 @@ section#debugger-io.debugger-jwt
             include ../token-editor-algorithms.pug
 
       include ../token-editor-common.pug
+      .debugger-footer.container
+        .row
+          .col-xs-6.website-validation
+            include ../token-editor-validation.pug
+          .col-xs-6.website-share
+            button.btn
+              span SHARE JWT
